[#261] BC Break: move toggle crud factories to crud package
